+++ hypervisor/Cargo.toml
@@ -23,7 +23,6 @@ mshv-ioctls = { git = "https://github.com/rust-vmm/mshv", branch = "main", optio
 serde = { version = "1.0.137", features = ["rc"] }
 serde_derive = "1.0.137"
 serde_json = "1.0.81"
-vm-device = { path = "../vm-device" }
 vm-memory = { version = "0.7.0", features = ["backend-mmap", "backend-atomic"] }
 vmm-sys-util = { version = "0.9.0", features = ["with-serde"] }
